refactor(api): move deployment logic to separate package
