+++ router/deployment.go
@@ -8,7 +8,7 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/go-vela/server/router/middleware/org"
 
-	"github.com/go-vela/server/api"
+	"github.com/go-vela/server/api/deployment"
 	"github.com/go-vela/server/router/middleware/perm"
 	"github.com/go-vela/server/router/middleware/repo"
 )
@@ -23,8 +23,8 @@ func DeploymentHandlers(base *gin.RouterGroup) {
 	// Deployments endpoints
 	deployments := base.Group("/deployments/:org/:repo", org.Establish(), repo.Establish())
 	{
-		deployments.POST("", perm.MustWrite(), api.CreateDeployment)
-		deployments.GET("", perm.MustRead(), api.GetDeployments)
-		deployments.GET("/:deployment", perm.MustRead(), api.GetDeployment)
+		deployments.POST("", perm.MustWrite(), deployment.CreateDeployment)
+		deployments.GET("", perm.MustRead(), deployment.ListDeployments)
+		deployments.GET("/:deployment", perm.MustRead(), deployment.GetDeployment)
 	} // end of deployments endpoints
 }
